requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-snowflake
snowflake-connector-python
snowflake-snowpark-python
boto3
python-dotenv
requests
dbt-core==1.9.0b3
dbt-snowflake==1.8.4
sqlfluff==3.0.6
astronomer-cosmos
psycopg2-binary
virtualenv